+++ pythran/pythonic/numpy/rank.hpp
@@ -11,7 +11,7 @@ PYTHONIC_NS_BEGIN
 namespace numpy
 {
   template <class E>
-  size_t rank(E const &expr)
+  long rank(E const &expr)
   {
     return E::value;
   }
